NCPlots.plot — Function

fig,ax,plt = plot(ds)

Plots dataset ds

fig,ax,plt = plot(var)

Plots variable var on the sphere. Returns fig ax and plt.

NCPlots.plotvar! — Function

plotvar!(ax,var)

Plots variables var into axis ax

NCPlots.addequator! — Function

addequator!(ax; lat=0, kwargs...)

Adds equator to axis ax

NCPlots.addmeridian! — Function

addmeridian!(ax; lon=0, kwargs...)

Adds meridian at lon=0 to axis ax

NCPlots.lonlat2xyz — Function

x,y,z = lonlat2xyz(lons,lats)

Convert lons, lats to x,y,z coordinates.

lons, lats can be either Vector or Matrix (for irregular grids in LAM models )

NCPlots.lonpadview — Function

lonpadview(data::Matrix) 
lonpadview(data::Vector)

Returns a view of data with the first row (assumed to be the longitude dimensions repeated at the end.

NCPlots.isperiodiclon — Function

isperiodiclon(lons)

Returns true if lons is a Vector and lons is periodic. false if lons is a Matrix (irregular grid LAM model)